Fix testcase so that if the initial file delete does work (which happens on non-windows systems) then the testcase doesn't fail
git-svn-id: https://svn.apache.org/repos/asf/tuscany/sca-java-2.x/trunk@1166222 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/domain-node/src/test/java/org/apache/tuscany/sca/runtime/ClassLoaderReleaseTestCase.java
@@ -39,10 +39,13 @@ public class ClassLoaderReleaseTestCase {
     public void testInstallDeployable() throws IOException, ActivationException{
         File f = copyFile("src/test/resources/sample-helloworld.jar");
         Node node = TuscanyRuntime.runComposite(null, f.toURI().toURL().toString());
-        Assert.assertFalse(f.delete());
-        Map<String, List<String>> scuris = node.getStartedCompositeURIs();
-        node.stopCompositeAndUninstallUnused(scuris.keySet().iterator().next(), scuris.get(scuris.keySet().iterator().next()).get(0));
-        Assert.assertTrue(f.delete());
+        // this delete should fail on Windows as the jar is locked, but the doesn't seem to happen in 'nix
+        boolean b = f.delete();
+        if (!b) {
+            Map<String, List<String>> scuris = node.getStartedCompositeURIs();
+            node.stopCompositeAndUninstallUnused(scuris.keySet().iterator().next(), scuris.get(scuris.keySet().iterator().next()).get(0));
+            Assert.assertTrue(f.delete());
+        }
     }
     
     private File copyFile(String fileName) throws IOException {
